Update 6 files

+++ update.js
@@ -17,10 +17,25 @@ module.exports = {
       message: "git pull"
     }
   }, {
+    when: "{{exists('app/face-animation')}}",
+    method: "shell.run",
+    params: {
+      path: "app/face-animation",
+      message: "git pull"
+    }
+  }, {
     method: "shell.run",
     params: {
       venv: "env",
       path: "app",
+      message: "uv pip install -r requirements.txt"
+    }
+  }, {
+    when: "{{exists('app/face-animation')}}",
+    method: "shell.run",
+    params: {
+      venv: "env",
+      path: "app/face-animation",
       message: "uv pip install -r requirements.txt"
     }
   }]
